+++ CHANGES.rst
@@ -18,6 +18,7 @@ Upcoming (unreleased)
 Added
 ~~~~~
 - Unified keyboard control for map, pointcloud, and video players
+- Support for subdirectories within site resources and exception if resource is not found
 - Filter for datasets without specific tag in frontend (EE)
 - Userdata annotation editor in web frontend (EE)
 - Option to calculate hashes for ingest and upload in parallel (EE)